Reset Computer Account Password
Without this commit if the computer account password become out of sync
with the domain, domain authentication stops flat. This can happen when
machine snapshots are rolled back for example.

This commit adds a second exec resource to reset the machine password if
it is detected to be out of sync.