Add 'fjoinoption' as a parameter
Allow passing a custom "FJoinOption" as a parameter.
See http://msdn.microsoft.com/en-us/library/aa392154(v=vs.85).aspx

Previous behavior was '1' unless the 'force' parameter was set to true,
which set the fjoinoption to '32'.  Parameterizing this allows for a
custom fjoinoption, which is sometimes necessary.

The default value is left at '1' and the 'force' parameter's behavior is
not changed by this. 'force' set to true will set the fjoinoption to 32,
even if a different value is passed with the fjoinoption parameter.