Remove `false` parameter in favor of `fjoinoption`
Upon further review, it seems there is an actual need for the ability to
set the `fjoinoption` argument of the `JoinDomainOrWorkgroup` method
exposed by the `Win32_ComputerSystem` WMI class. The fjoinoptions
arguments are actually bit flags which require the flexability to be
passed in a number of different combinations.

This commit eliminated the `force` parameter in favor of passing the raw
bit options. The new default value of fjoinoptions will be 1, which is a
standard domain join operation. If force functionality is still needed,
the proper bit option, 32, can be passed.