Update puppetlabs/reboot dependency
As per issue #9, the version of '0.9.1' for the dependency of the
`puppetlabs/reboot` module is incorrect. This was never a release and is
suspected to have originally been a typo. In addition, since the last
release of this module, much newer and stable releases of the
`puppetlabs/reboot` dependency have been made available.

This commit updates the dependency to require at least version 1.0.0 of
the `puppetlabs/reboot` module.